booking form bug -fix

+++ src/main/java/com/asg/shipping/bookingFormSH/dto/BookingFormUpdateDTO.java
@@ -1,6 +1,5 @@
 package com.asg.shipping.bookingFormSH.dto;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -34,10 +33,10 @@ public class BookingFormUpdateDTO {
 	private Long linePoid;
 	private Long salesmanPoid;
 	private Long comodityPoid;
-	private BigDecimal totalVolume;
-	private BigDecimal totalWeight;
+	private Double totalVolume;
+	private Double totalWeight;
 	private String unitPack;
-	private BigDecimal totalNoOfPacks;
+	private Double totalNoOfPacks;
 	private Long placeOfRecieptPoid;
 	private Long placeOfDelieveryPoid;
 	private Long portOfLoadingPoid;
